fixed resetPasswordPage access, fixed orders check before rendering in profileOrdersPage
