+++ src/pages/reset-password.tsx
@@ -3,17 +3,21 @@ import {
   Input,
   Button,
 } from "@ya.praktikum/react-developer-burger-ui-components";
-import { Link, Redirect } from "react-router-dom";
-import { resetPassword } from "../services/actions/profile";
+import { Link, Redirect, useHistory, useLocation } from "react-router-dom";
+import { resetPassword } from "../services/actions/user";
 
 import styles from "./login.module.css";
 import { useDispatch, useSelector } from "../utils/store-type";
 import { TForm } from "../utils/types";
 
 export function ResetPasswordPage() {
+  const history = useHistory();
   const dispatch = useDispatch();
+
+  const resetAllowed = useSelector((state) => state.auth.resetAllowed);
+  const reseted = useSelector((state) => state.auth.reseted);
+
   const [form, setValue] = useState<TForm>({ password: "", token: "" });
-  const resetAllowed = useSelector((state) => state.auth);
 
   const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setValue({ ...form, [e.target.name]: e.target.value });
@@ -31,6 +35,10 @@ export function ResetPasswordPage() {
 
   if (!resetAllowed) {
     return <Redirect to={{ pathname: "/forgot-password" }} />;
+  }
+
+  if (resetAllowed && reseted) {
+    history.replace({ pathname: "/login" });
   }
 
   return (
