+++ src/services/actions/user.ts
@@ -1,6 +1,14 @@
 import {
   AUTH_CHECKED,
   AUTH_FAILED,
+  EDIT_FAILED,
+  EDIT_REQUEST,
+  EDIT_SUCCESS,
+  RESET_FAILED,
+  RESET_REQUEST,
+  RESET_REQUEST_FAILED,
+  RESET_SUCCESS,
+  RESETED,
   SET_USER_REQUEST,
   SET_USER_SUCCESS,
 } from "../constants/profile";
@@ -8,6 +16,9 @@ import {
   refreshToken,
   BASE_URL,
   fetchWithRefresh,
+  editProfileRequest,
+  passwordResetRequest,
+  passwordReset,
 } from "../../utils/api";
 import { getCookie, setCookie } from "../../utils/utils";
 import { TForm } from "../../utils/types";
@@ -32,7 +43,7 @@ export const authFailedAcion = (): TAuthFailedAction => ({
   type: AUTH_FAILED,
 });
 
-// setting user section
+// Setting User
 
 export type TUserAction = {
   readonly type: typeof SET_USER_REQUEST;
@@ -51,30 +62,101 @@ export type TUserActions =
   | TAuthCheckAction
   | TAuthFailedAction;
 
+// Edit, Passwords Types
+
+export type TEditAction = {
+  readonly type: typeof EDIT_REQUEST;
+};
+
+export type TEditSuccessAction = {
+  readonly type: typeof EDIT_SUCCESS;
+  readonly payload?: TForm;
+};
+
+export type TEditFailedAction = {
+  readonly type: typeof EDIT_FAILED;
+};
+
+export type TResetAction = {
+  readonly type: typeof RESET_REQUEST;
+};
+
+export type TResetFailedRequestAction = {
+  readonly type: typeof RESET_REQUEST_FAILED;
+};
+
+export type TResetSuccessAction = {
+  readonly type: typeof RESET_SUCCESS;
+};
+
+export type TResetFailedAction = {
+  readonly type: typeof RESET_FAILED;
+};
+
+export type TResetedAction = {
+  readonly type: typeof RESETED;
+};
+
+export type TProfileActions =
+  | TEditAction
+  | TEditSuccessAction
+  | TEditFailedAction
+  | TResetAction
+  | TResetedAction
+  | TResetFailedAction
+  | TResetSuccessAction
+  | TResetFailedRequestAction;
+
+// ACTIONS
+
+// Setting user
+
 export const setUserAction = (): TUserAction => ({
   type: SET_USER_REQUEST,
 });
 
-export const setUserSuccessAction = (
-  user: TForm,
-): TUserSuccessAction => ({
+export const setUserSuccessAction = (user: TForm): TUserSuccessAction => ({
   type: SET_USER_SUCCESS,
   user,
 });
 
-// functions
+// Edit, Reset
 
-// export const checkUserAuth = (): AppThunk => (dispatch) => {
-//   if (getCookie("accessToken")) {
-//     dispatch(getUser())
-//       .finally(() => {
-//         dispatch({ type: AUTH_CHECKED, authChecked: true });
-//       })
-//       .catch((err: string) => {
-//         dispatch({ type: AUTH_FAILED, payload: err });
-//       });
-//   }
-// };
+export const editAction = (): TEditAction => ({
+  type: EDIT_REQUEST,
+});
+
+export const editSuccessAction = (): TEditSuccessAction => ({
+  type: EDIT_SUCCESS,
+});
+
+export const editFailedAction = (): TEditFailedAction => ({
+  type: EDIT_FAILED,
+});
+
+export const resetAction = (): TResetAction => ({
+  type: RESET_REQUEST,
+});
+
+export const resetFailedRequestAction = (): TResetFailedRequestAction => ({
+  type: RESET_REQUEST_FAILED,
+});
+
+export const resetSuccesAction = (): TResetSuccessAction => ({
+  type: RESET_SUCCESS,
+});
+
+export const resetFailedAction = (): TResetFailedAction => ({
+  type: RESET_FAILED,
+});
+
+export const resetSuccessData = (): TResetedAction => ({
+  type: RESETED,
+});
+
+// THUNKS
+
+// Token, Get User
 
 export interface ITokens {
   refreshToken: string;
@@ -101,16 +183,6 @@ export const updateToken = (): AppThunk => {
   };
 };
 
-interface IUser {
-  name: string;
-  email: string;
-}
-
-interface IGetUser {
-  user: IUser;
-  res: null;
-}
-
 export function getUser(): AppThunk {
   return function (dispatch) {
     fetchWithRefresh(`${BASE_URL}auth/user`, {
@@ -123,7 +195,53 @@ export function getUser(): AppThunk {
         dispatch(setUserSuccessAction(res.user));
       })
       .catch((error) => {
-        console.log('Error get user', error);
+        console.log("Error get user", error);
       });
   };
 }
+
+// Edit Profile
+// Forgot Passwod
+// Reset Password
+
+export const editProfile = (data: TForm): AppThunk => {
+  return function (dispatch) {
+    dispatch({ type: EDIT_REQUEST });
+    editProfileRequest(data)
+      .then((res) => {
+        dispatch({ type: EDIT_SUCCESS, payload: res.user });
+      })
+      .catch((err) => {
+        dispatch({ type: EDIT_FAILED, payload: err });
+      });
+  };
+};
+
+export const forgotPassword = (data: TForm): AppThunk => {
+  return function (dispatch) {
+    dispatch({ type: RESET_REQUEST });
+    passwordResetRequest(data)
+      .then((res) => {
+        dispatch({
+          type: RESET_SUCCESS,
+          payload: res.message,
+        });
+      })
+      .catch((err) => {
+        dispatch({ type: RESET_REQUEST_FAILED });
+        console.log(err);
+      });
+  };
+};
+
+export const resetPassword = (data: TForm): AppThunk => {
+  return function (dispatch) {
+    passwordReset(data)
+      .then((res) => {
+        dispatch({ type: RESETED });
+      })
+      .catch((err) => {
+        dispatch({ type: RESET_FAILED, payload: err });
+      });
+  };
+};
