+++ src/services/constants/profile.ts
@@ -8,6 +8,7 @@ export const RESET_REQUEST_FAILED: "RESET_REQUEST_FAILED" =
 
 export const RESET_SUCCESS: "RESET_SUCCESS" = "RESET_SUCCESS";
 export const RESET_FAILED: "RESET_FAILED" = "RESET_FAILED";
+export const RESETED: "RESETED" = "RESETED" 
 
 export const TOKEN_UPDATE: "TOKEN_UPDATE" = "TOKEN_UPDATE";
 export const TOKEN_UPDATE_FAILED: "TOKEN_UPDATE_FAILED" = "TOKEN_UPDATE_FAILED";
