+++ src/services/reducers/index.ts
@@ -5,7 +5,6 @@ import { ingredientReducer } from "./ingredientReducer";
 import { combineReducers } from "redux";
 import { constructorReducer } from "./constructorReducer";
 import { authReducer } from "./authReducers";
-import { userReducer } from "./userReducer";
 import { wsReducer } from "./wsReducer";
 
 
@@ -15,7 +14,6 @@ export const reducer = combineReducers({
   ingredient: ingredientReducer,
   cart: constructorReducer,
   auth: authReducer,
-  // user: userReducer,
   ws: wsReducer,
   wsProfile: wsProfileReducer
 });
