+++ src/utils/api.ts
@@ -197,7 +197,7 @@ export const fetchWithRefresh = async (
 export const passwordResetRequest = async (
   form: TForm
 ): Promise<TResponseBody<"user", TForm>> => {
-  return await request(`${BASE_URL}reset-password`, {
+  return await fetch(`${BASE_URL}password-reset`, {
     method: "POST",
     mode: "cors",
     cache: "no-cache",
@@ -208,7 +208,7 @@ export const passwordResetRequest = async (
     redirect: "follow",
     referrerPolicy: "no-referrer",
     body: JSON.stringify(form),
-  });
+  }).then(checkResponse)
 };
 
 export const passwordReset = async (
